% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coastal_geometry.R
\name{download_coastal_geometry}
\alias{download_coastal_geometry}
\title{Download 'TWDB' Coastal Geographic Boundaries}
\usage{
download_coastal_geometry(type)
}
\arguments{
\item{type}{character, required. One of \code{c("basin", "bay", "estuary",
"sub_watershed", "watershed")}}
}
\value{
sf
}
\description{
Download 'TWDB' Coastal Geographic Boundaries
}
