% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coastal_site_parameters.R
\name{download_coastal_site_parameters}
\alias{download_coastal_site_parameters}
\title{Download 'TWDB' Coastal Station Parameters}
\usage{
download_coastal_site_parameters(station)
}
\arguments{
\item{station}{character, required.}
}
\value{
tibble
}
\description{
Download 'TWDB' Coastal Station Parameters
}
