% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coastal_sites.R
\name{download_coastal_sites}
\alias{download_coastal_sites}
\title{Download 'TWDB' Coastal Station Information}
\usage{
download_coastal_sites(all_stations = FALSE, opts = list())
}
\arguments{
\item{all_stations}{logical. Defaults is FALSE. If FALSE, returns all
stations for which actual data is available. If TRUE, returns all stations
regardless of data availability.}

\item{opts}{list of curl options passed to crul::HttpClient()}
}
\value{
tibble
}
\description{
Download 'TWDB' Coastal Station Information
}
