% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_reservoir.R
\name{download_reservoir}
\alias{download_reservoir}
\title{Download "TWDB" Reservoir Data}
\usage{
download_reservoir(aggregate_by = NULL, region_name = NULL,
  reservoir = NULL, period = "historical")
}
\arguments{
\item{aggregate_by}{optional character. One of \code{c("statewide", "planning
region", "basin", "municipal", "climate region")}. Defaults to \code{NULL}.}

\item{region_name}{optional character. Documentation comming soon.}

\item{reservoir}{optional character. Documentation soon.}

\item{period}{optional character. Documentation soon}
}
\value{
dataframe
}
\description{
Download "TWDB" Reservoir Data
}
