% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_well.R
\name{download_well}
\alias{download_well}
\title{Download "TWDB" Individual Well Data}
\usage{
download_well(state_well_nmbr, opts = list())
}
\arguments{
\item{state_well_nmbr}{required character. State well number.}

\item{opts}{list of curl options passed to crul::HttpClient()}
}
\value{
dataframe
}
\description{
Download "TWDB" Individual Well Data
}
