% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coastal_geometry.R
\name{download_coastal_geometry}
\alias{download_coastal_geometry}
\title{Download 'TWDB' Coastal Geographic Boundaries}
\usage{
download_coastal_geometry(type, opts = list())
}
\arguments{
\item{type}{character, required. One of \code{c("basin", "bay", "estuary",
"sub_watershed", "watershed")}}

\item{opts}{list of curl options passed to crul::HttpClient()}
}
\value{
simple features data_frame
}
\description{
Download 'TWDB' Coastal Geographic Boundaries
}
