% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coastal_hydrology.R
\name{download_coastal_hydrology}
\alias{download_coastal_hydrology}
\title{Download 'TWDB' Historical Bay and Estuary Freshwater Inflow Data}
\usage{
download_coastal_hydrology(geoid, resample = NULL)
}
\arguments{
\item{geoid}{character, required.}

\item{resample}{character, optional. By default, data is returned in daily
interval. Specify one of \code{c("month", "year")} to get monthly or yearly
summed values.}
}
\description{
Download 'TWDB' Historical Bay and Estuary Freshwater Inflow Data
}
