% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_coastal_site_data.R
\name{download_coastal_site_data}
\alias{download_coastal_site_data}
\title{Download 'TWDB' Coastal Measurement Data}
\usage{
download_coastal_site_data(station, parameter, start_date, end_date,
  bin = "day")
}
\arguments{
\item{station}{character, required.}

\item{parameter}{character, required.}

\item{start_date}{character, required. Format must be \code{yyyy-mm-dd}.}

\item{end_date}{character, required. Format must be \code{yyyy-mm-dd}.}

\item{bin}{character. One of \code{"mon", "day", "hour"}. Default is
\code{"day"}}
}
\value{
tibble
}
\description{
Download 'TWDB' Coastal Measurement Data
}
\examples{
download_coastal_site_data(station = "MIDG", parameter =
  "seawater_salinity", start_date = "2010-01-01", end_date = "2010-12-31",
  bin = "hour")
}
